Allow restart of Gerrit instances
Only initialization of the environment was possible. This change
allows the restart of the Gerrit instances to simulate, for example,
an upgrade.

Feature: Issue 10594

+++ dockerised_local_env/gerrit-2/docker-entrypoint.sh
@@ -1,19 +1,16 @@
 #!/bin/bash -ex
 
-java -Xmx100g -jar /var/gerrit/bin/gerrit.war init -d /var/gerrit --batch --dev --no-auto-start --install-all-plugins
-#
-echo "Gerrit configuration:"
-cat /var/gerrit/etc/gerrit.config
-echo "Replication plugin configuration:"
-cat /var/gerrit/etc/replication.config
-
-echo "Remove git repos created during init phase"
-rm -fr /var/gerrit/git/*
-
 echo "Starting git daemon"
 /usr/local/bin/git-daemon.sh &
 
-echo "Waiting for initial replication"
-sleep 120
+if [[ $INIT == 1 ]]; then
+  java -Xmx100g -jar /var/gerrit/bin/gerrit.war init -d /var/gerrit --batch --dev --no-auto-start --install-all-plugins
+
+  echo "Remove git repos created during init phase"
+  rm -fr /var/gerrit/git/*
+
+  echo "Waiting for initial replication"
+  sleep 120
+fi
 
 java -Xmx100g -jar /var/gerrit/bin/gerrit.war daemon
